Force Flushing of File Buffers
We noticed that, for the 3D data, the plot was not updating smoothly,
but in bunches. It seemed like buffering and it was.  So, we forcibly
write the buffer out after some writes so maintain consistency.

This is a compromise.  This probably slows down the high data rates
in 1D, but we haven't had much problem with that, lately, on new
hardware.  But, the data rate for 3D is low enough, that it takes a
few seconds to fill.  So, this make a big difference in how the
3D plotting appears to function.