Explicit Buffering, Flushing
This patch is related to the previous one in which we manually flush
the incoming data files after writing.  Originally, we tried to
force synchronous writes when we open the files, but that didn't
work.  I forget the exact reason, but we proved that it didn't
work as expected.  However, we left the explicit code for setting
the buffering at the time of open() so that we would know that it
was on purpose and so that someone else doesn't try synchronous
to solve some other problem.